datasets operations: move them to example based fixup
